Split bootstrap.go into six smaller, more focussed files
